pyproject.toml
[project]
name = "echotrace"
version = "0.1.0"
description = "Branchable, version-controlled AI conversation system built with Streamlit."
authors = [{ name = "thomas yee", email = "tyeetale@gmail.com" }]
license = "MIT"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "streamlit>=1.30",
    "openai>=1.11",
    "networkx>=3.2",
    "pyvis>=0.3.2",
    "python-dotenv>=1.0.1"
]

[project.optional-dependencies]
dev = ["black>=24.0", "isort>=5.12"]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
